refactor(FOROME-871): move DrawerTable & DrawerPreView to separate components
